instantiate synthesis dataframe

instantiate synthesis dataframe

synth_df(USUBJID, n_new = 10, width = 3, maxweight = 2/3)

Arguments

USUBJID
character vector of person identifiers

n_new
number of new persons to generate, defaults to 10

width
number of persons from original trial to use for new person synthesis, defaults to 3

maxweight
maximum allowed weight for one person for synthesis of new person, defaults to 2/3

Value

dataframe relating new persons to source persons and their their contribution's weight

Examples

synth_df(USUBJID = letters)

#>    synid USUBJID     weight selected
#> 1      1       m 0.18033877    FALSE
#> 2      1       w 0.21689988    FALSE
#> 3      1       l 0.68016292     TRUE
#> 4      2       e 0.49884561     TRUE
#> 5      2       f 0.64167935    FALSE
#> 6      2       o 0.66028435    FALSE
#> 7      3       o 0.09602416    FALSE
#> 8      3       d 0.76560016    FALSE
#> 9      3       f 0.76967480     TRUE
#> 10     4       i 0.99071231    FALSE
#> 11     4       e 0.97052090     TRUE
#> 12     4       y 0.38918276    FALSE
#> 13     5       x 0.46118646    FALSE
#> 14     5       o 0.31524175     TRUE
#> 15     5       m 0.17467589    FALSE
#> 16     6       b 0.53157354    FALSE
#> 17     6       w 0.49363702    FALSE
#> 18     6       k 0.77930863     TRUE
#> 19     7       k 0.35420680    FALSE
#> 20     7       e 0.82519942     TRUE
#> 21     7       u 0.27381825    FALSE
#> 22     8       l 0.57004495     TRUE
#> 23     8       b 0.33571908    FALSE
#> 24     8       f 0.59626279    FALSE
#> 25     9       d 0.19151803     TRUE
#> 26     9       p 0.94776394    FALSE
#> 27     9       b 0.54248041    FALSE
#> 28    10       c 0.54460339     TRUE
#> 29    10       y 0.27859715    FALSE
#> 30    10       v 0.44670247    FALSE