create a synthetic table from syndf and an original SDTM dataframe

create a synthetic table from syndf and an original SDTM dataframe

synthesize(syndf, df, cat_fuzz = 1)

Arguments

syndf
synthesis dataframe

df
original SDTM dataframe

cat_fuzz
fuzz factor for noise on categorical variables, defaults to 1

Value

synthesized SDTM dataframe

Examples

synthesize(synth_df(CRC305ABC_DM$USUBJID), CRC305ABC_DM)

#> # A tibble: 10 x 18
#>    synid RFSTDTC             RFENDTC             RFXSTDTC           
#>    <int> <dttm>              <dttm>              <dttm>             
#>  1     1 2013-08-24 00:03:00 2013-09-21 00:00:00 2013-08-24 00:03:00
#>  2     2 2013-09-07 16:20:00 2014-01-21 16:00:00 2013-09-07 16:20:00
#>  3     3 2013-09-27 16:20:20 2013-12-19 08:00:00 2013-09-27 16:20:20
#>  4     4 2013-08-11 00:20:00 2013-10-26 16:00:00 2013-08-11 00:20:00
#>  5     5 2013-08-09 16:22:00 2013-12-20 16:00:00 2013-08-09 16:22:00
#>  6     6 2013-07-02 16:25:20 2013-11-04 08:00:00 2013-07-02 16:25:20
#>  7     7 2013-07-12 16:20:00 2013-11-18 00:00:00 2013-07-12 16:20:00
#>  8     8 2013-06-17 08:30:00 2013-10-17 16:00:00 2013-06-17 08:30:00
#>  9     9 2013-04-18 16:25:00 2013-07-05 16:00:00 2013-04-18 16:25:00
#> 10    10 2013-08-14 00:20:00 2013-10-30 08:00:00 2013-08-14 00:20:00
#> # … with 14 more variables: RFXENDTC <dttm>, RFICDTC <dttm>, RFPENDTC <dttm>,
#> #   AGE <dbl>, STUDYID <chr>, DOMAIN <chr>, SITEID <chr>, AGEU <chr>,
#> #   SEX <chr>, RACE <chr>, ETHNIC <chr>, ARMCD <chr>, ARM <chr>, COUNTRY <chr>